refactor: Made changes request in PR40
