+++ package.json
@@ -31,7 +31,6 @@
     "inversify-binding-decorators": "^4.0.0",
     "inversify-express-utils": "^6.3.2",
     "nunjucks": "^3.2.0",
-    "redis": "^3.0.2",
     "reflect-metadata": "^0.1.13",
     "tslib": "^1.11.0"
   },
