+++ src/utils/ConfigLoader.ts
@@ -47,18 +47,3 @@ export const getExpressAppConfig = (directory: string) => (app: express.Applicat
     app.locals.ROOT_URI = ROOT_URI;
 };
 
-export function returnEnvVarible(name: string, defaultVal?: string): string {
-
-  const variable = process.env[name];
-
-  if (!variable) {
-    if (defaultVal !== undefined) {
-      return defaultVal;
-    }
-    throw Error(`Variable ${name} was not found on env files.`);
-  }
-
-  return variable;
-}
-
-
