+++ test/controllers/OtherReasonController.test.ts
@@ -10,39 +10,38 @@ import { OTHER_REASON_PAGE_URI } from '../../src/utils/Paths';
 import { OK, UNPROCESSABLE_ENTITY, MOVED_TEMPORARILY } from 'http-status-codes';
 import { SessionStore, SessionMiddleware, CookieConfig } from 'ch-node-session';
 import { Redis } from 'ioredis';
-import Substitute from '@fluffy-spoon/substitute';
+import Substitute, { Arg } from '@fluffy-spoon/substitute';
 import { RequestHandler } from 'express';
 import * as CookieUtil from 'ch-node-session/lib/utils/CookieUtils';
+import { returnEnvVarible } from '../../src/utils/EnvironmentUtils';
+import { Session } from 'ch-node-session/lib/session/model/Session';
+import { wrapValue } from 'ch-node-session/lib/utils/EitherAsyncUtils';
 const pageHeading = 'Tell us why you’re appealing this penalty';
 const errorSummaryHeading = 'There is a problem with the information you entered';
 const invalidTitleErrorMessage = 'You must give your reason a title';
 const invalidDescriptionErrorMessage = 'You must give us more information';
 
 
-const config: CookieConfig = {
-    cookieName: '__SID',
-    cookieSecret: 'S+CmgW/ivLEaiFTzm87a1cyH+ZbD81yukx8n7e/efzQ='
-};
+const app = (sessionStore: SessionStore) => createApplication(container => {
+
+    const config: CookieConfig = {
+        cookieName: returnEnvVarible('COOKIE_NAME'),
+        cookieSecret: returnEnvVarible('COOKIE_SECRET')
+    };
+
+    const sessionHandler = SessionMiddleware(config, sessionStore);
+    setupFakeAuth(container);
+    container.bind<RequestHandler>(SessionMiddleware).toConstantValue(sessionHandler);
+    container.bind(SessionStore).toConstantValue(sessionStore);
+
+});
 
 describe('OtherReasonController', () => {
 
     describe('GET request', () => {
         it('should return 200 response', async () => {
-            const app = createApplication(container => {
 
-                const redis = {
-                    get: (s: string) => Promise.resolve('OK')
-                } as Redis;
-
-                const sessionStore = new SessionStore(redis);
-                const sessionHandler = SessionMiddleware(config, sessionStore);
-                setupFakeAuth(container);
-                container.bind<RequestHandler>(SessionMiddleware).toConstantValue(sessionHandler);
-                container.bind(SessionStore).toConstantValue(sessionStore);
-
-            });
-
-            await request(app).get(OTHER_REASON_PAGE_URI)
+            await request(app(Substitute.for<SessionStore>())).get(OTHER_REASON_PAGE_URI)
                 .expect(response => {
                     expect(response.status).to.be.equal(OK);
                     expect(response.text).to.include(pageHeading)
@@ -53,21 +52,8 @@ describe('OtherReasonController', () => {
 
     describe('POST request', () => {
         it('should return 422 response with rendered error messages when invalid data was submitted', async () => {
-            const app = createApplication(container => {
 
-                const redis = {
-                    get: (s: string) => Promise.resolve('OK')
-                } as Redis;
-
-                const sessionStore = new SessionStore(redis);
-                const sessionHandler = SessionMiddleware(config, sessionStore);
-                setupFakeAuth(container);
-                container.bind<RequestHandler>(SessionMiddleware).toConstantValue(sessionHandler);
-                container.bind(SessionStore).toConstantValue(sessionStore);
-
-            });
-
-            await request(app).post(OTHER_REASON_PAGE_URI)
+            await request(app(Substitute.for<SessionStore>())).post(OTHER_REASON_PAGE_URI)
                 .send({})
                 .expect(response => {
                     expect(response.status).to.be.equal(UNPROCESSABLE_ENTITY);
@@ -82,18 +68,10 @@ describe('OtherReasonController', () => {
             const title = 'Some title';
             const description = 'Some description';
 
-            const app = createApplication(container => {
+            const sessionStore = Substitute.for<SessionStore>();
+            sessionStore.load(Arg.any()).returns(wrapValue({ title, description }));
 
-                const redis = Substitute.for<Redis>();
-                redis.get('session::other-reason').returns(Promise.resolve(JSON.stringify({ title, description })));
-                const sessionStore = new SessionStore(redis);
-                const sessionHandler = SessionMiddleware(config, sessionStore);
-                setupFakeAuth(container);
-                container.bind<RequestHandler>(SessionMiddleware).toConstantValue(sessionHandler);
-                container.bind(SessionStore).toConstantValue(sessionStore);
-
-            });
-            await request(app).post(OTHER_REASON_PAGE_URI)
+            await request(app(sessionStore)).post(OTHER_REASON_PAGE_URI)
                 .send({ title, description })
                 .expect(response => {
                     expect(response.status).to.be.equal(OK);
